PMM-6433 Fix environment variables passing.
